+++ Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/Textures.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3816933984354201bb7f9c0424c97f79
+timeCreated: 1680361438
